edit taskcontroller and viewfiles

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,18 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data=[];
+        if(\Auth::check()){
+            $user=\Auth::user();
+            $tasks=$user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data=[
+                'user'=>$user,
+                'tasks'=>$tasks,
+            ];
+        }
         
-        return view('tasks.index',[
-            'tasks'=>$tasks,    
-        ]);
+        return view('tasks.index',$data);
     }
 
     /**
@@ -49,10 +56,10 @@ class TasksController extends Controller
            'status'=>'required|max:10',
         ]);
         
-        $task=new Task;
-        $task->content=$request->content;
-        $task->status=$request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+           'content'=>$request->content,
+           'status'=>$request->status,
+        ]);
         
         return redirect('/');
     }
@@ -67,7 +74,11 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view('tasks.show',['task'=>$task,]);
+        if(\Auth::id()==$task->user_id){
+            return view('tasks.show',['task'=>$task,]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -80,7 +91,11 @@ class TasksController extends Controller
     {
         $task=Task::find($id);
         
-        return view('tasks.edit',['task'=>$task]);
+        if(\Auth::id()==$task->user_id){
+            return view('tasks.edit',['task'=>$task,]);
+        }else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -98,9 +113,11 @@ class TasksController extends Controller
         ]);
         
         $task=Task::find($id);
-        $task->status=$request->status;
-        $task->content=$request->content;
-        $task->save();
+        if(\Auth::id()==$task->user_id){
+            $task->content=$request->content;
+            $task->status=$request->status;
+            $task->save();
+        }
         
         return redirect('/');
     }
@@ -114,7 +131,10 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task=Task::find($id);
-        $task->delete();
+        
+        if(\Auth::id()==$task->user_id){
+            $task->delete();
+        }
         
         return redirect('/');
     }
